+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 ####10.06.2016
 * [ADDED] New Helperfunction table.concat
 * [ADDED] MySql.helper.update()
+* [ADDED] MySql.helper.delete()
 * [REMOVED] Senseless Admincommands
 * [REMOVED] Old Function MySQL_GetString removed
 * [REMOVED] Old Function MySQL_SetVar removed
 * [REMOVED] Old Function MySQL_SetString removed
+* [REMOVED] Old Function MySQL_DelRow removed
 * [REMOVED] MySql.helper.getColumn doesn't make any sense
 * [CHANGED] Rename MySql.helper.getFieldValueSync to MySql.helper.getValueSync
 * [CHANGED] Function MySql.helper.getFieldValueSync supports now other compare options then "=" (set a table {comparer, value} instead of value in this case)